+++ models/FieldPaletteModel.php
@@ -44,80 +44,85 @@ class FieldPaletteModel extends \Model
     /**
      * Find all published fieldpalette elements by their ids
      *
-     * @param array $arrIds     An array of fielpalette ids
+     * @param array $arrIds An array of fielpalette ids
      * @param array $arrOptions An optional options array
+     * @param array $columns Additional clauses columns
+     * @param array $values Additional clauses values
      *
      * @return \Model\Collection|\ContentModel|null A collection of models or null if there are no fieldpalette elements
      */
-    public static function findPublishedByIds(array $arrIds = [], array $arrOptions = [])
+    public static function findPublishedByIds(array $arrIds = [], array $arrOptions = [], array $columns = [], array $values = null)
     {
         $t = static::$strTable;
 
-        if (!is_array($arrIds) || empty($arrIds))
-        {
+        if (!is_array($arrIds) || empty($arrIds)) {
             return null;
         }
 
         $arrColumns = ["$t.id IN(" . implode(',', array_map('intval', $arrIds)) . ")"];
 
-        if (!BE_USER_LOGGED_IN)
-        {
+        if (!BE_USER_LOGGED_IN) {
             $time         = \Date::floorToMinute();
             $arrColumns[] = "($t.start='' OR $t.start<='$time') AND ($t.stop='' OR $t.stop>'" . ($time + 60) . "') AND $t.published='1'";
         }
 
-        if (!isset($arrOptions['order']))
-        {
+        if (!isset($arrOptions['order'])) {
             $arrOptions['order'] = "$t.sorting";
         }
 
-        return static::findBy($arrColumns, null, $arrOptions);
+        $arrColumns = array_merge($arrColumns, $columns);
+
+        return static::findBy($arrColumns, $values, $arrOptions);
     }
 
     /**
      * Find all published fieldpalette elements by their parent ID and parent table
      *
-     * @param integer $intPid         The article ID
-     * @param string  $strParentTable The parent table name
-     * @param string  $strParentField The parent field name
-     * @param array   $arrOptions     An optional options array
+     * @param integer $intPid The article ID
+     * @param string $strParentTable The parent table name
+     * @param string $strParentField The parent field name
+     * @param array $arrOptions An optional options array
+     * @param array $columns Additional clauses columns
+     * @param array $values Additional clauses values
      *
      * @return \Model\Collection|\ContentModel|null A collection of models or null if there are no fieldpalette elements
      */
-    public static function findPublishedByPidAndTableAndField($intPid, $strParentTable, $strParentField, array $arrOptions = [])
+    public static function findPublishedByPidAndTableAndField($intPid, $strParentTable, $strParentField, array $arrOptions = [], array $columns = [], array $values = [])
     {
         $t = static::$strTable;
 
         $arrColumns = ["$t.pid=? AND $t.ptable=? AND $t.pfield=?"];
 
-        if (!BE_USER_LOGGED_IN)
-        {
+        if (!BE_USER_LOGGED_IN) {
             $time         = \Date::floorToMinute();
             $arrColumns[] = "($t.start='' OR $t.start<='$time') AND ($t.stop='' OR $t.stop>'" . ($time + 60) . "') AND $t.published='1'";
         }
 
-        if (!isset($arrOptions['order']))
-        {
+        if (!isset($arrOptions['order'])) {
             $arrOptions['order'] = "$t.sorting";
         }
 
-        return static::findBy($arrColumns, [$intPid, $strParentTable, $strParentField], $arrOptions);
+        $arrColumns = array_merge($arrColumns, $columns);
+        $values     = array_merge([$intPid, $strParentTable, $strParentField], $values);
+
+        return static::findBy($arrColumns, $values, $arrOptions);
     }
 
     /**
      * Find all published fieldpalette elements by their parent ID and parent table
      *
-     * @param array  $arrPids        The parent ids
+     * @param array $arrPids The parent ids
      * @param string $strParentTable The parent table name
      * @param string $strParentField The parent field name
-     * @param array  $arrOptions     An optional options array
+     * @param array $arrOptions An optional options array
+     * @param array $columns Additional clauses columns
+     * @param array $values Additional clauses values
      *
      * @return \Model\Collection|\ContentModel|null A collection of models or null if there are no fieldpalette elements
      */
-    public static function findPublishedByPidsAndTableAndField($arrPids, $strParentTable, $strParentField, array $arrOptions = [])
+    public static function findPublishedByPidsAndTableAndField($arrPids, $strParentTable, $strParentField, array $arrOptions = [], array $columns = [], array $values = [])
     {
-        if (!is_array($arrPids) || empty($arrPids))
-        {
+        if (!is_array($arrPids) || empty($arrPids)) {
             return null;
         }
 
@@ -127,42 +132,47 @@ class FieldPaletteModel extends \Model
 
         $arrColumns[] = "$t.ptable=? AND $t.pfield=?";
 
-        if (!BE_USER_LOGGED_IN)
-        {
+        if (!BE_USER_LOGGED_IN) {
             $time         = \Date::floorToMinute();
             $arrColumns[] = "($t.start='' OR $t.start<='$time') AND ($t.stop='' OR $t.stop>'" . ($time + 60) . "') AND $t.published='1'";
         }
 
-        if (!isset($arrOptions['order']))
-        {
+        if (!isset($arrOptions['order'])) {
             $arrOptions['order'] = "FIELD($t.pid, " . implode(',', array_map('intval', $arrPids)) . "), $t.sorting";
         }
 
-        return static::findBy($arrColumns, [$strParentTable, $strParentField], $arrOptions);
+        $arrColumns = array_merge($arrColumns, $columns);
+        $values     = array_merge([$strParentTable, $strParentField], $values);
+
+        return static::findBy($arrColumns, $values, $arrOptions);
     }
 
 
     /**
      * Find all fieldpalette elements by their parent ID and parent table
      *
-     * @param integer $intPid         The article ID
-     * @param string  $strParentTable The parent table name
-     * @param string  $strParentField The parent field name
-     * @param array   $arrOptions     An optional options array
+     * @param integer $intPid The article ID
+     * @param string $strParentTable The parent table name
+     * @param string $strParentField The parent field name
+     * @param array $arrOptions An optional options array
+     * @param array $columns Additional clauses columns
+     * @param array $values Additional clauses values
      *
      * @return \Model\Collection|\ContentModel|null A collection of models or null if there are no fieldpalette elements
      */
-    public static function findByPidAndTableAndField($intPid, $strParentTable, $strParentField, array $arrOptions = [])
+    public static function findByPidAndTableAndField($intPid, $strParentTable, $strParentField, array $arrOptions = [], array $columns = [], array $values = [])
     {
         $t = static::$strTable;
 
         $arrColumns = ["$t.pid=? AND $t.ptable=? AND $t.pfield=?"];
 
-        if (!isset($arrOptions['order']))
-        {
+        if (!isset($arrOptions['order'])) {
             $arrOptions['order'] = "$t.sorting";
         }
 
-        return static::findBy($arrColumns, [$intPid, $strParentTable, $strParentField], $arrOptions);
+        $arrColumns = array_merge($arrColumns, $columns);
+        $values     = array_merge([$intPid, $strParentTable, $strParentField], $values);
+
+        return static::findBy($arrColumns, $values, $arrOptions);
     }
 }